Add files via upload
Updates for this morning.  Got a 'good' csv output -- but the KF with IMU data is "linearized" -- a real goofy output.  Corrected chart looks better.  Graph axis still doesn't have 42.0 it shows 4.20 E+01 and the titles don't show up.  I'd like to add a 0.001 pad to the x-axis and y-axis.  Not sure how to format the x- and y-axis or force the titles in the plot_first_dim routine.